+++ CHANGELOG.md
@@ -7,7 +7,8 @@ and this project adheres to [Semantic Versioning](http://semver.org).
 ## [Unreleased] - TBD
 ### Added
 - `GetEventSubscriptionsListAsync` method with enhanced filtering and sorting capabilities for retrieving event subscriptions
-- `GetEventSubscriptionByIdAsync` method for retrieving event subscription details by ID using the v2 event-subscriptions endpoint
+- `GetEventSubscriptionAsync` method for retrieving event subscription details by ID using the v2 event-subscriptions endpoint
+- `DeleteEventSubscriptionAsync` method for deleting event subscription
 
 ### Changed
 - Upgraded .NET Core runtime to .NET 7.0 for Tests and Examples projects